remove registry-like logic from metrics manager
